Updated create-update

+++ library/functions.py
@@ -38,14 +38,21 @@ def update_order(api_key, dto):
 
     existing_item = monday_board.get_items_by_column_values('text5', str(dto['id']), return_items_as='json')\
         .get('data').get('items_page_by_column_values').get('items')[0]
-    subitems_ids = [x['id'] for x in existing_item.get('subitems', [])]
     existing_item_id = int(existing_item['id'])
 
-    monday_board.change_multiple_column_values({
+    is_cancel = dto.get('is_cancel', False)
+
+    column_values = {'status': 'בוטל'} if is_cancel else {
         'short_text': dto['note'],  # Note
         'dropdown': dto['location'],  # Location
-        **({'status': 'בוטל'} if dto['is_cancel'] else {})  # Order Status
-    }, existing_item_id)
+    }
+
+    monday_board.change_multiple_column_values(column_values, existing_item_id)
+
+    if is_cancel:
+        return
+
+    subitems_ids = [x['id'] for x in existing_item.get('subitems', [])]
 
     for item_id in subitems_ids:
         monday_board.delete_item(item_id)
